Split weight-free export code into focused checkpoint, spec-building,   runtime, and export orchestration modules. Rename checkpoint transform   helpers to checkpoint_transforms.py to align with existing transform   naming patterns.
  Make checkpoint preparation conditional, avoid mutating source checkpoint
  directories during .bin to safetensors conversion, and remove script-level
  profiler imports from exporter code.
